nav bar updates based on if user is logged in or not without having to refresh page
