+++ src/reducers/index.js
@@ -1,0 +1,10 @@
+
+// const initialState = {
+//     isAuth: false
+// }
+
+// export const reducer = (state = initialState, action) => {
+//     switch (action.type) {
+//         case()
+//     }
+// }
